requirements.txt
streamlit
crewai
langchain-community-tools
requests
python-dotenv
matplotlib
vaderSentiment
langchain
langchain-groq
duckduckgo-search
numpy
